Update rate feature and change jwt duration

+++ src/main/java/OOSE_Final_Project/Blog/dto/req/RateReq.java
@@ -15,5 +15,5 @@ public class RateReq {
     long blogId;
 
     @NonNull
-    long rateStar;
+    double rateStar;
 }
