+++ src/main/java/OOSE_Final_Project/Blog/dto/res/RateRes.java
@@ -13,7 +13,7 @@ public class RateRes {
 
     long rating;
 
-    long rateStar;
+    double rateStar;
 
     long id;
 }
